+++ services/serverupdate/model_update_update_schedule_payload.go
@@ -22,8 +22,7 @@ type UpdateUpdateSchedulePayload struct {
 	// REQUIRED
 	Enabled *bool `json:"enabled"`
 	// Can be cast to int32 without loss of precision.
-	// REQUIRED
-	Id *int64 `json:"id"`
+	Id *int64 `json:"id,omitempty"`
 	// Can be cast to int32 without loss of precision.
 	// REQUIRED
 	MaintenanceWindow *int64 `json:"maintenanceWindow"`
@@ -39,10 +38,9 @@ type _UpdateUpdateSchedulePayload UpdateUpdateSchedulePayload
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
 // will change when the set of required properties is changed
-func NewUpdateUpdateSchedulePayload(enabled *bool, id *int64, maintenanceWindow *int64, name *string, rrule *string) *UpdateUpdateSchedulePayload {
+func NewUpdateUpdateSchedulePayload(enabled *bool, maintenanceWindow *int64, name *string, rrule *string) *UpdateUpdateSchedulePayload {
 	this := UpdateUpdateSchedulePayload{}
 	this.Enabled = enabled
-	this.Id = id
 	this.MaintenanceWindow = maintenanceWindow
 	this.Name = name
 	this.Rrule = rrule
@@ -81,26 +79,34 @@ func (o *UpdateUpdateSchedulePayload) SetEnabled(v *bool) {
 	o.Enabled = v
 }
 
-// GetId returns the Id field value
+// GetId returns the Id field value if set, zero value otherwise.
 func (o *UpdateUpdateSchedulePayload) GetId() *int64 {
-	if o == nil {
+	if o == nil || IsNil(o.Id) {
 		var ret *int64
 		return ret
 	}
-
 	return o.Id
 }
 
-// GetIdOk returns a tuple with the Id field value
+// GetIdOk returns a tuple with the Id field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *UpdateUpdateSchedulePayload) GetIdOk() (*int64, bool) {
-	if o == nil {
+	if o == nil || IsNil(o.Id) {
 		return nil, false
 	}
 	return o.Id, true
 }
 
-// SetId sets field value
+// HasId returns a boolean if a field has been set.
+func (o *UpdateUpdateSchedulePayload) HasId() bool {
+	if o != nil && !IsNil(o.Id) {
+		return true
+	}
+
+	return false
+}
+
+// SetId gets a reference to the given int64 and assigns it to the Id field.
 func (o *UpdateUpdateSchedulePayload) SetId(v *int64) {
 	o.Id = v
 }
@@ -180,7 +186,9 @@ func (o *UpdateUpdateSchedulePayload) SetRrule(v *string) {
 func (o UpdateUpdateSchedulePayload) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
 	toSerialize["enabled"] = o.Enabled
-	toSerialize["id"] = o.Id
+	if !IsNil(o.Id) {
+		toSerialize["id"] = o.Id
+	}
 	toSerialize["maintenanceWindow"] = o.MaintenanceWindow
 	toSerialize["name"] = o.Name
 	toSerialize["rrule"] = o.Rrule